change damage from float to int

+++ Assets/Script/Data/Damage.cs
@@ -2,7 +2,7 @@ namespace Script.Data
 {
     public struct Damage
     {
-        public float Amount;
+        public int Amount;
         public DamageType Type;
     }
 }
